v4.0

+++ app/Models/Attachment.php
@@ -2,11 +2,15 @@
 
 namespace App\Models;
 
+use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\HtmlString;
+use function PHPUnit\Framework\isNull;
 
 class Attachment extends Model
 {
@@ -21,9 +25,7 @@ class Attachment extends Model
         'order_id',
         'payment_id',
         'payment_request_id',
-        'doc_id',
-        'order_request_id',
-        'logistic_id',
+        'proforma_invoice_id',
     ];
 
     protected $casts = [
@@ -39,16 +41,18 @@ class Attachment extends Model
 
     protected static function booted()
     {
-        static::creating(function ($post) {
-            $post->user_id = auth()->id() ?? null;
+        static::creating(function ($attachment) {
+            $attachment->user_id = auth()->id() ?? null;
         });
 
+
         static::deleting(function ($attachment) {
-            if ($attachment->file_path && File::exists(public_path($attachment->file_path))) {
+            if (!$attachment->isUsedElsewhere() && $attachment->file_path && File::exists(public_path($attachment->file_path))) {
                 Storage::disk('public')->delete($attachment->file_path);
             }
         });
     }
+
 
     /**
      * Get the user that owns the attachment.
@@ -75,14 +79,6 @@ class Attachment extends Model
     }
 
     /**
-     * Get the logistic associated with the attachment.
-     */
-    public function logistic()
-    {
-        return $this->belongsTo(Logistic::class);
-    }
-
-    /**
      * Get the payment request associated with the attachment.
      */
     public function paymentRequest()
@@ -91,19 +87,31 @@ class Attachment extends Model
     }
 
     /**
-     * Get the document associated with the attachment.
-     */
-    public function document()
-    {
-        return $this->belongsTo(Doc::class, 'doc_id');
-    }
-
-    /**
      * Get the order request associated with the attachment.
      */
-    public function orderRequest()
+    public function proformaInvoice()
     {
-        return $this->belongsTo(OrderRequest::class, 'order_request_id');
+        return $this->belongsTo(ProformaInvoice::class, 'proforma_invoice_id');
+    }
+
+    //Computational Methods
+
+    public function getCreatedAtBy()
+    {
+        $creator = optional($this->user)->fullName;
+        $time = $this->created_at->format('F d, Y');;
+        $message = "Uploaded on {$time} by {$creator}";
+
+        return new HtmlString("<span class='italic'>$message</span>");
+    }
+
+    public function isUsedElsewhere(): bool
+    {
+        return $this->where('file_path', $this->file_path)
+            ->where('name', $this->name)
+            ->where('id', '!=', $this->id)
+            ->whereNull('deleted_at')
+            ->exists();
     }
 
     public static function hasTitleContainingPart($title, $orderId)
@@ -111,5 +119,26 @@ class Attachment extends Model
         return self::where('order_id', $orderId)
             ->where('file_path', 'LIKE', "%{$title}%")
             ->exists();
+    }
+
+    public static function getRelatedRecords(Attachment $attachment, $relations)
+    {
+        if (empty($attachment->file_path) && empty($attachment->name)) {
+            return collect();
+        }
+
+        $query = self::query()
+            ->when(!empty($attachment->file_path), function ($query) use ($attachment) {
+                $query->where('file_path', $attachment->file_path);
+            })
+            ->when(!empty($attachment->name), function ($query) use ($attachment) {
+                $query->where('name', $attachment->name);
+            });
+
+        if (!empty($relations)) {
+            $query->with($relations);
+        }
+
+        return $query->get();
     }
 }
